[Solax] Changes in values parser for X3-Hybrid-G4 (#17549)
* Fixes in indexes from Ruepert #1

+++ bundles/org.openhab.binding.solax/src/main/java/org/openhab/binding/solax/internal/model/local/X1HybridG4Data.java
@@ -49,7 +49,7 @@ public class X1HybridG4Data extends CommonLocalDeviceData {
     }
 
     @Override
-    public short getFeedInPower() {
+    public int getFeedInPower() {
         return getFromRawData(32);
     }
